chore: migrate to 11ty (#199)
* chore: migrate to 11ty

* chore: update readme
