+++ .prettierrc.js
@@ -1,0 +1,7 @@
+module.exports = {
+  bracketSpacing: true,
+  printWidth: 120,
+  singleQuote: true,
+  semi: false,
+  trailingComma: 'es5',
+};
