+++ package.json
@@ -1,42 +1,31 @@
 {
-  "name": "bulma-admin-dashboard",
-  "description": "Free Admin Dashboard Template Build with Bulma.io",
-  "version": "1.0.0",
-  "author": "Irfan Maulana <mazipanneh@gmail.com>",
-  "private": false,
-  "license": "MIT",
-  "scripts": {
-    "dev": "cross-env NODE_ENV=development webpack-dev-server --hot --color --open",
-    "build": "rimraf ./dist && cross-env NODE_ENV=production webpack --color --progress --hide-modules",
-    "publish-demo": "node ./publish-demo.js",
-    "test": "echo \"Error: no test specified\" && exit 1"
-  },
-  "dependencies": {
-    "bulma-dracula": "^1.1.1"
-  },
-  "devDependencies": {
-    "@babel/core": "7.11.6",
-    "@babel/preset-env": "7.11.5",
-    "apexcharts": "3.21.0",
-    "babel-loader": "8.1.0",
-    "compression-webpack-plugin": "6.0.2",
-    "copy-webpack-plugin": "6.1.1",
-    "cross-env": "7.0.2",
-    "css-loader": "4.3.0",
-    "file-loader": "6.1.0",
-    "gh-pages": "3.1.0",
-    "html-webpack-plugin": "4.5.0",
-    "mini-css-extract-plugin": "0.11.2",
-    "node-sass": "4.14.1",
-    "optimize-css-assets-webpack-plugin": "5.0.4",
-    "purgecss-webpack-plugin": "3.0.0",
-    "rimraf": "3.0.2",
-    "sass-loader": "10.0.2",
-    "style-loader": "1.2.1",
-    "terser-webpack-plugin": "4.2.2",
-    "upath": "1.2.0",
-    "webpack": "4.44.2",
-    "webpack-cli": "3.3.12",
-    "webpack-dev-server": "3.11.0"
-  }
+	"name": "bulma-admin-dashboard",
+	"description": "Free Admin Dashboard Template Build with Bulma.io",
+	"version": "1.0.0",
+	"author": "Irfan Maulana <mazipanneh@gmail.com>",
+	"private": false,
+	"license": "MIT",
+	"scripts": {
+		"watch:sass": "node-sass --importer node_modules/node-sass-magic-importer/dist/cli.js --watch src/scss -o dist/css",
+		"build:sass": "node-sass --importer node_modules/node-sass-magic-importer/dist/cli.js src/scss -o dist/css",
+		"build-local:sass": "node-sass --importer node_modules/node-sass-magic-importer/dist/cli.js src/scss -o dist/css",
+		"build:eleventy": "NODE_ENV=production eleventy",
+		"watch:eleventy": "cross-env NODE_ENV=development eleventy --watch --serve",
+		"dev": "npm-run-all build-local:sass --parallel watch:*",
+		"build": "npm-run-all build:*",
+		"test": "echo \"Error: no test specified\" && exit 1"
+	},
+	"dependencies": {
+		"bulma": "^0.9.1",
+		"bulma-dracula": "^1.1.1"
+	},
+	"devDependencies": {
+		"@11ty/eleventy": "^0.11.0",
+		"node-sass": "^4.14.1",
+		"node-sass-magic-importer": "^5.3.2",
+		"cross-env": "7.0.2",
+		"npm-run-all": "^4.1.5",
+		"optimize-css-assets-webpack-plugin": "5.0.4",
+		"rimraf": "3.0.2"
+	}
 }
