+++ src/_data/data.js
@@ -1,0 +1,9 @@
+const isProd = process.env.NODE_ENV === 'production';
+const publicPath = isProd ? '/bulma-admin-dashboard-template' : '';
+
+module.exports = {
+  config: {
+    isProd,
+    publicPath,
+  },
+};
