+++ src/_includes/jsonLd.ejs
@@ -1,0 +1,73 @@
+<script type="application/ld+json">
+  {
+    "@context": "http://schema.org",
+    "@type": "WebSite",
+    "url": "https://www.mazipan.github.io/",
+    "name": "Irfan Maulana | Front End Developer",
+    "author": "Irfan Maulana",
+    "image": "http://mazipan.github.io/images/irfan-maulana.jpg",
+    "description": "Irfan Maulana is Front End Developer from Indonesia - Man that craft some code to build a beauty and readable code, experienced in web and desktop technology.",
+    "sameAs": [
+      "https://www.facebook.com/mazipanneh",
+      "https://instagram.com/maz_ipan",
+      "https://twitter.com/Maz_Ipan",
+      "https://id.linkedin.com/in/irfanmaulanamazipan",
+      "https://www.slideshare.net/IrfanMaulana21",
+      "https://github.com/mazipan"
+    ]
+  }
+</script>
+<script type="application/ld+json">
+  {
+    "@context": "http://schema.org",
+    "@type": "Person",
+    "email": "mailto:mazipanneh@gmail.com",
+    "image": "http://mazipan.github.io/images/irfan-maulana.jpg",
+    "jobTitle": "Software Engineer",
+    "name": "Irfan Maulana",
+    "url": "https://www.mazipan.github.io/",
+    "sameAs": [
+      "https://www.facebook.com/mazipanneh",
+      "https://instagram.com/maz_ipan",
+      "https://twitter.com/Maz_Ipan",
+      "https://id.linkedin.com/in/irfanmaulanamazipan",
+      "https://www.slideshare.net/IrfanMaulana21",
+      "https://github.com/mazipan"
+    ]
+  }
+</script>
+<script type="application/ld+json">
+  {
+    "@context": "http://schema.org",
+    "@type": "BreadcrumbList",
+    "itemListElement": [
+      {
+        "@type": "ListItem",
+        "position": 1,
+        "item": {
+          "@id": "http://mazipan.github.io/",
+          "name": "Home",
+          "image": "http://mazipan.github.io/images/irfan-maulana.jpg"
+        }
+      },
+      {
+        "@type": "ListItem",
+        "position": 2,
+        "item": {
+          "@id": "http://mazipan.github.io/demo/",
+          "name": "Demo",
+          "image": "http://mazipan.github.io/images/irfan-maulana.jpg"
+        }
+      },
+      {
+        "@type": "ListItem",
+        "position": 3,
+        "item": {
+          "@id": "http://mazipan.github.io/bulma-admin-dashboard-template",
+          "name": "bulma-admin-dashboard-template",
+          "image": "https://mazipan.github.io/bulma-admin-dashboard-template/logo.png"
+        }
+      }
+    ]
+  }
+</script>
