+++ src/_includes/metaTags.ejs
@@ -1,0 +1,16 @@
+<meta name="robots" content="index, follow" />
+<meta http-equiv="Content-Type" content="text/html; charset=UTF-8" />
+<meta http-equiv="X-UA-Compatible" content="IE=edge" />
+<meta name="viewport" content="width=device-width,initial-scale=1,minimum-scale=1,maximum-scale=7">
+
+<meta name="language" content="en-EN" />
+<meta name="author" content="Irfan Maulana" />
+<title>Free Admin Dashboard Template Build with Bulma.io By Irfan Maulana</title>
+
+<link rel="stylesheet" href="https://maxcdn.bootstrapcdn.com/font-awesome/4.7.0/css/font-awesome.min.css">
+<link rel="stylesheet" href="<%= data.config.publicPath %>/css/main.css" />
+<script async src="<%= data.config.publicPath %>/js/module-event.js"></script>
+
+<!-- Global site tag (gtag.js) - Google Analytics -->
+<script async src="https://www.googletagmanager.com/gtag/js?id=UA-25065548-2"></script>
+<script>window.dataLayer = window.dataLayer || []; function gtag() { dataLayer.push(arguments); } gtag('js', new Date()); gtag('config', 'UA-25065548-2');</script>
