--- src/main.js
@@ -1,4 +1,0 @@
-require('./main.scss')
-
-require('./module-event.js')
-require('./module-chart.js')
