+++ src/scss/main.scss
@@ -5,10 +5,11 @@ $selected: #6272a4;
 $text: #f8f8f2;
 $cyan: #8be9fd;
 
-.nav-tag{
+.nav-tag {
   padding-right: 30px;
 }
-.tag-notif{
+
+.tag-notif {
   border-radius: 50%;
   position: absolute;
   top: 5px;
@@ -18,110 +19,119 @@ $cyan: #8be9fd;
     margin-left: -2px;
   }
 }
-.hero{
+
+.hero {
   background-color: $theme;
 }
-.navbar{
+
+.navbar {
   background: $theme;
 
-  &-item{
+  &-item {
     background-color: $theme;
     color: $text !important;
-    &.is-active{
+
+    &.is-active {
       background-color: $selected !important;
       color: $text !important;
     }
+
     &:hover {
       background-color: $selected !important;
       color: $text !important;
     }
   }
 
-  .is--brand{
+  .is--brand {
     &:hover {
       background-color: $theme !important;
     }
   }
 
-  &-brand-logo{
+  &-brand-logo {
     max-height: 3rem !important;
     background-color: $theme;
     border-radius: .25em;
     padding: .25em;
   }
 
-  &-dropdown{
+  &-dropdown {
     background-color: $theme;
-    .navbar-item{
+
+    .navbar-item {
       background-color: $theme;
     }
   }
 
-  &-burger{
+  &-burger {
     background-color: $theme !important;
     border: 0;
-    span{
+
+    span {
       background-color: $cyan;
     }
   }
 
-  &-link{
+  &-link {
     color: $text;
 
-    figure{
-      img{
+    figure {
+      img {
         border-radius: 50%;
         max-height: 2em;
       }
     }
 
-    &:hover{
+    &:hover {
       background-color: $theme;
     }
   }
 }
 
-.message.is-dark{
-  .message-body{
+.message.is-dark {
+  .message-body {
     background-color: lighten($theme, 5%);
   }
 }
 
-.breadcrumb:not(:last-child){
+.breadcrumb:not(:last-child) {
   margin-bottom: 0;
 }
 
 @media screen and (min-width: 1088px) {
-  .navbar-item.has-dropdown:hover .navbar-link, .navbar-item.has-dropdown.is-active .navbar-link {
+
+  .navbar-item.has-dropdown:hover .navbar-link,
+  .navbar-item.has-dropdown.is-active .navbar-link {
     background-color: $theme;
   }
 }
+
 // navigation left area
-.aside{
+.aside {
   background: $theme;
   padding: 2em .2em 0 1em;
 
-  a{
+  a {
     color: $text;
   }
 
   .menu-list {
     li {
       margin-top: .5em;
-      
+
     }
 
     .has-children {
       max-height: 36px;
       overflow: hidden;
-      transition: 0.6s max-height ease-out ;
+      transition: 0.6s max-height ease-out;
       position: relative;
-      
-      > a {
+
+      >a {
         position: relative;
       }
 
-      > a::before {
+      >a::before {
         content: "\f0d7";
         font-family: "FontAwesome";
         transition: .2s ease-out transform;
@@ -137,7 +147,7 @@ $cyan: #8be9fd;
     .has-children.open {
       max-height: 1000px;
 
-      > a::before {
+      >a::before {
         transform: rotate(-180deg) translateY(40%);
       }
     }
@@ -145,15 +155,15 @@ $cyan: #8be9fd;
 }
 
 // main area
-.main{
+.main {
   padding: 2em 1em 0 1em;
 }
 
-.menu-label{
+.menu-label {
   color: $cyan;
   margin-left: .5em;
 }
 
-.block{
+.block {
   margin-top: .5em;
 }
